refactor(studio): split repository dto contracts
